requirements.txt
streamlit==1.27.2
qrcode==7.4.2
pillow==10.0.0 